+++ core/class/mybin.class.php
@@ -60,14 +60,10 @@ class mybin extends eqLogic {
         $cmd->save();
         /****************************************************************************************************************/
 
-        $month = 1 * date('n');
-        $week = 1 * date('W');
-        $day = 1 * date('w');
         $hour = 1 * date('G');
         $minute = 1 * date('i');
-        log::add(__CLASS__, 'debug', $this->getHumanName() . ' checkbin: day ' . $day . ', hour ' . $hour . ', minute ' . $minute);
-        $change = $change + $this->checkNotifBin($month, $week, $day, $hour, $minute);
-        $change = $change + $this->checkAckBin($month, $week, $day, $hour, $minute);
+        $change = $change + $this->checkNotifBin();
+        $change = $change + $this->checkAckBin();
         if ($change > 0 || ($hour == 0 && $minute == 5)) {
             $this->refreshWhole();
         }
@@ -84,12 +80,21 @@ class mybin extends eqLogic {
         }
     }
     
-    public function checkNotifBin($month, $week, $day, $hour, $minute) {
+    public function checkNotifBin() {
         if (!$this->getIsEnable()) {
             return 0;
         }
         
         $dt = new DateTime("now");
+        $delay = $this->getConfiguration('notif_days', 0);
+        if ($delay > 0) {
+            $dt->modify('+'.$delay.' day');
+        }
+        $month = 1 * date('n');
+        $week = 1 * date('W');
+        $day = 1 * date('w');
+        $hour = 1 * date('G');
+        $minute = 1 * date('i');
         
         $isSpecificDay = false;
         $isMonth = false;
@@ -97,11 +102,14 @@ class mybin extends eqLogic {
         $isday = false;
         $ishour = false;
         $isminute = false;
+
+        /*
         $myday = $day;
         $myweek = $week;
         $mymonth = $month;
-        if ($this->getConfiguration('notif_veille') == 1) {
-            $myday = $myday + 1;
+        $delay = $this->getConfiguration('notif_days', 0);
+        if ($delay > 0) {
+            $myday = $myday + $delay;
             if ($myday == 7) {
                 $myday = 0;
             }
@@ -122,8 +130,10 @@ class mybin extends eqLogic {
                 }
             }
 
-            $dt->modify('+1 day');
+            $dt->modify('+'.$delay.' day');
         }
+        */
+
         $specificDays = $this->getConfiguration('specific_day');
         if (is_array($specificDays)) {
             foreach ($specificDays as $specificDay) {
@@ -136,14 +146,14 @@ class mybin extends eqLogic {
                 }
             }
         }
-        if ($this->getConfiguration('month_'.$mymonth) == 1) {
+        if ($this->getConfiguration('month_'.$month) == 1) {
             $isMonth = true;
         }
-        if (($myweek%2 == 0 && $this->getConfiguration('paire') == 1) || ($myweek%2 != 0 && $this->getConfiguration('impaire') == 1)) {
+        if (($week%2 == 0 && $this->getConfiguration('paire') == 1) || ($week%2 != 0 && $this->getConfiguration('impaire') == 1)) {
             $isweek = true;
         }
         for ($i = 0; $i <= 6; $i++) {
-            if ($this->getConfiguration('day_'.$i) == 1 && $i == $myday) {
+            if ($this->getConfiguration('day_'.$i) == 1 && $i == $day) {
                 $isday = true;
                 break;
             }
@@ -163,12 +173,18 @@ class mybin extends eqLogic {
         }
     }
     
-    public function checkAckBin($month, $week, $day, $hour, $minute) {
+    public function checkAckBin() {
         if (!$this->getIsEnable()) {
             return 0;
         }
         
         $dt = new DateTime("now");
+
+        $month = 1 * date('n');
+        $week = 1 * date('W');
+        $day = 1 * date('w');
+        $hour = 1 * date('G');
+        $minute = 1 * date('i');
         
         $isSpecificDay = false;
         $ismonth = false;
@@ -267,15 +283,6 @@ class mybin extends eqLogic {
         $this->refreshWidget();
     }
 
-    public function lastWeekNumberOfYear() {
-        $year = date('Y');
-        $week_count = date('W', strtotime($year . '-12-31'));
-        if ($week_count == '01'){
-            $week_count = date('W', strtotime($year . '-12-24'));
-        }
-        return intval($week_count);
-    }
-
     // Fonction exécutée automatiquement avant la création de l'équipement
     public function preInsert() {
         if ($this->getConfiguration('type','') == 'whole') {
@@ -288,7 +295,7 @@ class mybin extends eqLogic {
             $this->setDisplay('width', '260px');
             $this->setConfiguration('hour', 8);
             $this->setConfiguration('minute', 0);
-            $this->setConfiguration('notif_veille', 1);
+            $this->setConfiguration('notif_days', 1);
             $this->setConfiguration('notif_hour', 20);
             $this->setConfiguration('notif_minute', 0);
             $this->setConfiguration('paire', 1);
@@ -307,7 +314,13 @@ class mybin extends eqLogic {
     //Fonction exécutée automatiquement avant la mise à jour de l'équipement
     public function preUpdate() {
         if ($this->getConfiguration('type','') <> 'whole') {
-            if ($this->getConfiguration('notif_veille') == 0) {
+            if ($this->getConfiguration('notif_days', '') <> '') {
+                $options = array('options' => array('min_range' => 0));
+                if (!filter_var($this->getConfiguration('notif_days'), FILTER_VALIDATE_INT, $options)) {
+                    throw new Exception(__('Le nombre de jours pour la notification doit être un entier positif ou être laissé vide',__FILE__));
+                }
+            }
+            if ($this->getConfiguration('notif_days', 0) == 0) {
                 if ($this->getConfiguration('notif_hour') > $this->getConfiguration('hour')) {
                     throw new Exception(__('L\'heure de notification est après l\'heure de collecte',__FILE__));
                 }
@@ -443,8 +456,8 @@ class mybin extends eqLogic {
                         continue;
                     }
                     $dtCheck = DateTime::createFromFormat("Y-m-d", $dtDisplay->format("Y-m-d"));
-                    if ($eqLogic->getConfiguration('notif_veille') == 1 && $calendarType == 'notif') {
-                        $dtCheck->modify('+1 day');
+                    if ($eqLogic->getConfiguration('notif_days', 0) > 0 && $calendarType == 'notif') {
+                        $dtCheck->modify('+'.$eqLogic->getConfiguration('notif_days').' day');
                     }
                     if ($eqLogic->checkIfBin($dtCheck)) {
                         $color = $eqLogic->getConfiguration('color');