+++ plugin_info/install.php
@@ -51,6 +51,9 @@ require_once dirname(__FILE__) . '/../../../core/php/core.inc.php';
             if ($eqLogic->getConfiguration('color') === 'braun') {
                 $eqLogic->setConfiguration('color', 'brown');
             }
+            if ($eqLogic->getConfiguration('notif_veille') == 1) {
+                $eqLogic->setConfiguration('notif_days', 1);
+            }
             
             $cmd = $eqLogic->getCmd(null, 'counter');
             if (!is_object($cmd)) {